fix model

+++ src/main/java/topjava/graduation/model/Dish.java
@@ -26,7 +26,7 @@ public class Dish extends AbstractNamedEntity {
 
     @NotNull
     @Column(name = "price", nullable = false)
-    private Double price;
+    private Integer price;
 
     @JsonBackReference(value = "restaurant")
     @JoinColumn(name = "restaurant_id", nullable = false)
@@ -34,14 +34,14 @@ public class Dish extends AbstractNamedEntity {
     @OnDelete(action = OnDeleteAction.CASCADE)
     private Restaurant restaurant;
 
-    public Dish(Long id, LocalDate date, String name, Double price, Restaurant restaurant) {
+    public Dish(Long id, LocalDate date, String name, Integer price, Restaurant restaurant) {
         super(id, name);
         this.date = date;
         this.price = price;
         this.restaurant = restaurant;
     }
 
-    public Dish(String name, Double price, Restaurant restaurant, LocalDate date) {
+    public Dish(String name, Integer price, Restaurant restaurant, LocalDate date) {
         super(name);
         this.date = date;
         this.price = price;
